tools/port_tester/requirements.txt
pyserial>=3.5
PySimpleGUI>=4.60.0
